test: Lotto 클래스 단위테스트 합침
